뿌리기 api 추가

+++ src/main/java/com/kakaopay/throwing/domain/Room.java
@@ -20,5 +20,5 @@ public class Room extends BaseEntity {
     private String room;
 
     @Column(name = "user")
-    private long user;
+    private int user;
 }
